refactor(logistics): fold the fleet inventory list into the logistics page
Fleet logistics split what the hangar keeps on one page: the inventory
panels sat behind a "View inventories" button while the ledger spanning
them stayed on the landing page. The panels, the empty state and the
create action now sit above the tables, matching /hangar/inventories.

The old path redirects rather than 404s, so links people already shared
keep working, and the detail crumbs lose the hop that would now land on
the page their logistics crumb already points at.
