+++ app/frontend/frontend/pages/fleets/[slug]/logistics/inventories/[inventory].vue
@@ -214,13 +214,6 @@ const crumbs = computed(() => [
     to: { name: "fleet-logistics", params: { slug: props.fleet.slug } },
     label: t("nav.fleets.logistics.index"),
   },
-  {
-    to: {
-      name: "fleet-logistics-inventories",
-      params: { slug: props.fleet.slug },
-    },
-    label: t("headlines.logistics.inventories"),
-  },
 ]);
 </script>
