+++ app/frontend/frontend/pages/fleets/[slug]/logistics/routes.ts
@@ -19,21 +19,11 @@ export const routes: RouteRecordRaw[] = [
     },
   },
   {
+    // The inventory list moved onto the logistics page itself. The path stays so
+    // links people already shared keep resolving.
     path: "inventories/",
     name: "fleet-logistics-inventories",
-    component: () =>
-      import("@/frontend/pages/fleets/[slug]/logistics/inventories/index.vue"),
-    meta: {
-      backgroundImage: "bg-8",
-      title: "fleets.logistics.inventories",
-      needsAuthentication: true,
-      access: [
-        "fleet:inventories:manage",
-        "fleet:inventories:create",
-        "fleet:manage",
-      ],
-      customTitle: true,
-    },
+    redirect: { name: "fleet-logistics" },
   },
   {
     path: "inventories/:inventory/",
