+++ app/frontend/translations/en/actions.json
@@ -250,7 +250,6 @@
         "hideRemoved": "Hide Removed"
       },
       "logistics": {
-        "viewInventories": "View Inventories",
         "viewRequests": "View Requests",
         "newRequest": "New Request",
         "createRequest": "Create Request",
