+++ app/frontend/translations/es/actions.json
@@ -235,7 +235,6 @@
         "hideRemoved": "Ocultar eliminados"
       },
       "logistics": {
-        "viewInventories": "View Inventories",
         "viewRequests": "View Requests",
         "newRequest": "New Request",
         "createRequest": "Create Request",
