+++ app/frontend/translations/fr/actions.json
@@ -235,7 +235,6 @@
         "hideRemoved": "Masquer les retirées"
       },
       "logistics": {
-        "viewInventories": "View Inventories",
         "viewRequests": "View Requests",
         "newRequest": "New Request",
         "createRequest": "Create Request",
