+++ app/frontend/translations/zh-TW/actions.json
@@ -235,7 +235,6 @@
         "hideRemoved": "Hide Removed"
       },
       "logistics": {
-        "viewInventories": "View Inventories",
         "viewRequests": "View Requests",
         "newRequest": "New Request",
         "createRequest": "Create Request",
